ITS#9613, ITS#9614 - Document multival, idlexp, and maxentrysize options to slapd-mdb
multival, idlexp, and maxentrysize are all new options as of OpenLDAP 2.5 that were missing from the admin guide.  Add these options and additional information beyond what is found in the man pages as appropriate.

Also fixed some discrepencies between the slapd.conf and slapd-config admin guide sections on the Database backends as well as missing information in that section.