ITS#9917 Remove 'h' and 'p' from options[] in client tools
The options -h and -p got removed from client tools in commit
66af4cfd5d3c559fcdf6d722b02fa66416b95f01. However, they were still
present in the options[] array in several client tools source files. So,
if one of those tools got executed with -h or -p followed by a value,
this lead to the error "unrecognized option -", without mentioning
which option was problematic. Removing 'h' and 'p' from options[] fixes
this.