Update msuser.ldif
The 'memberOf' attribute is part of InetOrgPerson/Core schema of Openldap. The attribute of the same name causes issues in loading the slapd and thus this name needs to be modified. I made the change to be random MSADatmemberOf but the developers of this project can hopefully modify the code in a way that it works with openldap core functionality